structural changes, remove unused classes
