+++ src/test/scala/org/virtuslab/beholder/LambdaDslTests.scala
@@ -2,13 +2,21 @@ package org.virtuslab.beholder
 
 import java.sql.Date
 
+import org.virtuslab.beholder.filters.dsl.FilterFactory
 import org.virtuslab.beholder.filters.json.JsonFilterFields
 import org.virtuslab.beholder.filters.{ FilterAPI, FilterDefinition }
 import org.virtuslab.beholder.suites.FiltersTestSuite
 import org.virtuslab.beholder.filters.json.JsonFilterFields._
 import org.virtuslab.unicorn.LongUnicornPlay.driver.simple._
 
-class DslTests extends AppTest with FiltersTestSuite[Unit] {
+abstract class DslTest extends AppTest with FiltersTestSuite[Unit] {
+  override def doFilters(data: BaseFilterData, currentFilter: FilterDefinition): Seq[UserMachineViewRow] = {
+    import data._
+    filter.filter(currentFilter)
+  }
+}
+
+class LambdaDslTests extends DslTest {
   override def createFilter(data: BaseFilterData): FilterAPI[UserMachineViewRow, Unit] = {
     import org.virtuslab.beholder.filters.dsl.DSL._
     create(usersMachinesQuery) {
@@ -20,14 +28,9 @@ class DslTests extends AppTest with FiltersTestSuite[Unit] {
           "capacity" as JsonFilterFields.ignore[Option[BigDecimal]] from machine.capacity
     }.mapped(UserMachineViewRow.tupled)
   }
-
-  override def doFilters(data: BaseFilterData, currentFilter: FilterDefinition): Seq[UserMachineViewRow] = {
-    import data._
-    filter.filterWithTotalEntitiesNumber(currentFilter).content
-  }
 }
 
-class DslTests2 extends AppTest with FiltersTestSuite[Unit] {
+class PartialFunctionDslTest extends DslTest {
   override def createFilter(data: BaseFilterData): FilterAPI[UserMachineViewRow, Unit] = {
     import org.virtuslab.beholder.filters.dsl.DSL._
     create(usersMachinesQuery) {
@@ -40,8 +43,32 @@ class DslTests2 extends AppTest with FiltersTestSuite[Unit] {
     }.mapped(UserMachineViewRow.tupled)
   }
 
-  override def doFilters(data: BaseFilterData, currentFilter: FilterDefinition): Seq[UserMachineViewRow] = {
-    import data._
-    filter.filterWithTotalEntitiesNumber(currentFilter).content
+}
+
+class FilterFactoryTests extends DslTest {
+  def createFilter(data: BaseFilterData): FilterAPI[UserMachineViewRow, Unit] = {
+
+    val q = usersMachinesQuery.map {
+      case (user, machine) =>
+        (user.email, machine.system, machine.cores, machine.created, machine.capacity)
+    }
+
+    val names = Seq(
+      "email",
+      "system",
+      "cores",
+      "created",
+      "capacity"
+    )
+
+    val fields = Seq(
+      inText,
+      inText,
+      inIntField,
+      inRange(inField[Date]("date")),
+      JsonFilterFields.ignore[Option[BigDecimal]]
+    )
+
+    FilterFactory.crate(q, fields, names).mapped(UserMachineViewRow.tupled)
   }
 }
